+++ init.rb
@@ -56,9 +56,9 @@ Tog::Plugins.helpers CoreHelper
 
 Tog::Interface.sections(:admin).add "Home", "/admin/dashboard"
 
-Tog::Interface.sections(:admin).add "Manage", "/admin/configuration"
-Tog::Interface.sections(:admin).tabs("Manage").add_item "Configuration", "/admin/configuration"
-Tog::Interface.sections(:admin).tabs("Manage").add_item "Abuses", "/admin/manage_abuses"
+manage = Tog::Interface.sections(:admin).add "Manage", "/admin/configuration"
+manage.add_item "Configuration", "/admin/configuration"
+manage.add_item "Abuses", "/admin/manage_abuses"
 
 Tog::Interface.sections(:member).add "Home", "/"
 Tog::Interface.sections(:site).add "Home", "/"