+++ test/test_helper.rb
@@ -2,6 +2,7 @@ ENV["RAILS_ENV"] = "test"
 require File.expand_path(File.dirname(__FILE__) + "/../../../../config/environment")
 require 'test_help'
 require 'test/unit'
+require 'action_view/test_case'
 require 'mocha'
 require 'shoulda'